refactor: move TestsBase to building blocks
