--- src/2-Services/Identity/Tests/Unit/Domain/TestData/TestsBase.cs
@@ -1,32 +1,0 @@
-﻿
-namespace TaskoMask.Services.Identity.Domain.Tests.Unit.TestData
-{
-    public abstract class TestsBase:IDisposable
-    {
-
-        /// <summary>
-        /// Run before each test method
-        /// </summary>
-        public TestsBase()
-        {
-            FixtureSetup();
-        }
-
-
-
-        /// <summary>
-        /// Each test class should setup its fixture
-        /// </summary>
-        protected abstract void FixtureSetup();
-
-
-
-        /// <summary>
-        /// Run after each test method
-        /// </summary>
-        public void Dispose()
-        {
-
-        }
-    }
-}
