+++ src/2-Services/Monolith/Tests/Unit/Domain.Tests.Unit/Membership/OperatorUnitTests.cs
@@ -2,14 +2,14 @@
 using System;
 using TaskoMask.BuildingBlocks.Domain.Exceptions;
 using TaskoMask.BuildingBlocks.Contracts.Resources;
-using TaskoMask.Services.Monolith.Domain.Tests.Unit.TestData;
 using TaskoMask.Services.Monolith.Domain.Tests.Unit.TestData.ObjectMothers;
 using TaskoMask.Services.Monolith.Domain.DomainModel.Membership.Entities;
 using Xunit;
+using TaskoMask.BuildingBlocks.Test;
 
 namespace TaskoMask.Services.Monolith.Domain.Tests.Unit.Membership
 {
-    public class OperatorUnitTests : TestsBase
+    public class OperatorUnitTests : UnitTestsBase
     {
 
         [Fact]
